app/requirements.txt
scikit-learn==0.24.1
numpy
scipy
pillow
